refactor the media handler
